automat-v2/spiralredeem_communicator/requirements.txt
paho-mqtt>=1.4.0
pyserial>=3.4

